Update contributions and user:count algorithms, add extraction algorithm
