+++ ohsomeTools/proc/data_aggregation/elements_ratio_aggregation.py
@@ -21,13 +21,18 @@ from qgis.core import (
     QgsProcessingParameterBoolean,
     QgsProcessingParameterDateTime,
     QgsWkbTypes,
+    QgsProcessing
 )
 
-from ohsomeTools.common import AGGREGATION_SPECS
-from ..procDialog import run_processing_alg
+from qgis.utils import iface
+
+from ohsomeTools.utils import configmanager
+
+from ohsomeTools.common import AGGREGATION_SPECS, client
+from ohsomeTools.proc.procDialog import run_processing_alg
 
 
-class ElementsPerimeter(QgsProcessingAlgorithm):
+class ElementsRatioAggregation(QgsProcessingAlgorithm):
     """
     This is an example algorithm that takes a vector layer and
     creates a new identical one.
@@ -49,6 +54,7 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
     PARAMETER = "PARAMETER"
     INPUT = "INPUT"
     FILTER = "FILTER"
+    FILTER_2 = "FILTER_2"
     check_activate_temporal = "check_activate_temporal"
     check_show_metadata = "check_show_metadata"
     timeout_input = "timeout_input"
@@ -62,12 +68,25 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
     MONTHS = "MONTHS"
     DAYS = "DAYS"
     RADIUS = "RADIUS"
+    GROUPBY = "GROUPBY"
     check_keep_geometryless = "check_keep_geometryless"
     check_merge_geometries = "check_merge_geometries"
     group_by_values_line_edit = "group_by_values_line_edit"
     group_by_key_line_edit = "group_by_key_line_edit"
     formats = ["json", "geojson"]
-    parameters = [i for i in AGGREGATION_SPECS["elements/perimeter"]]
+    parameters = [
+        i.split("/")[1] for i in AGGREGATION_SPECS.keys() if "elements" in i
+    ]
+    group_by = [
+        "",
+        "/boundary",
+        "/key",
+        "/tag",
+        "/type",
+        "/boundary/groupBy/tag",
+    ]
+    PERIOD = "PERIOD"
+    group_by_boundary = 'group_by_boundary'
 
     def tr(self, string):
         """
@@ -76,7 +95,7 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
         return QCoreApplication.translate("Processing", string)
 
     def createInstance(self):
-        return ElementsPerimeter()
+        return ElementsRatioAggregation()
 
     def name(self):
         """
@@ -86,14 +105,14 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
         lowercase alphanumeric characters only and no spaces or other
         formatting characters.
         """
-        return "dataaggregationelementsperimeter"
+        return "elementsratioaggregation"
 
     def displayName(self):
         """
         Returns the translated algorithm name, which should be used for any
         user-visible display of the algorithm name.
         """
-        return self.tr("Elements Perimeter")
+        return self.tr("Elements Ratio Aggregation")
 
     def group(self):
         """
@@ -118,28 +137,86 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
         should provide a basic description about what the algorithm does and the
         parameters and outputs associated with it.
         """
-        return self.tr("Example algorithm short description")
+        return self.tr(
+            """
+        <p>Aggregation endpoints for the <strong>Ohsome-API</strong>. See <a href="https://docs.ohsome.org/ohsome-api/v1/">documentation</a>. </p>
+        <p><strong>Parameters</strong></p>
+        <ul>
+        <li><em>Radius</em>: Radius for point layers.</li>
+        <li><em>Period</em>: ISO 8601 Period, eg. /P1M for a monthly aggregation.</li>
+        <li><em>Show Metadata</em>: Include metadata into the query response. Depending on the request of the request this can increase the response data size significantly.</li>
+        <li><em>Keep without geometry</em>: Some results don&#39;t contain geometries but metadata. Decide if you wan&#39;t to keep them or only return ones with geometries. If checked, the geometry less features will be stored separately.</li>
+        <li><em>Harmonize geometries</em>: Check this to <ins>automatically merge compatible geometry types</ins> It is recommended to keep this checked. The benefit is that the amount of written layers will be massively reduced. The reason is that results may contain single and multi-geometries at once (Polygon, MultiPolygon etc.) and without combining them one layer per geometry type will be written, resulting in an increased number of layers.</li>
+        <li><em>Qgis temporal feature</em>: Automatically enable the temporal feature for new layers where applicable. This is only applied to responses that contain geometries and in that manner only on those geometry layers it makes sense for.</li>
+        <li><em>Clip geometries</em>: Specify whether the returned geometries of the features should be clipped to the query’s spatial boundary. <ins>Ony available for the data extraction endpoints</ins></li>
+        </ul>
+        """
+        )
 
     def initAlgorithm(self, config=None):
         """
         Here we define the inputs and output of the algorithm, along
         with some other properties.
         """
+        # get dates from metadata
+        provider = configmanager.read_config()["providers"][0]
+        clnt = client.Client(provider)
+        metadata = clnt.check_api_metadata(iface)
+
+        start_date_string = (
+            metadata.get("extractRegion")
+            .get("temporalExtent")
+            .get("fromTimestamp")
+        )
+
+        end_date_string = (
+            metadata.get("extractRegion")
+            .get("temporalExtent")
+            .get("toTimestamp")
+        )
 
         # We add the input vector features source. It can have any kind of
         # geometry.
         self.addParameter(
             QgsProcessingParameterVectorLayer(
-                self.LAYER, self.tr("Query Layer")
+                self.LAYER,
+                self.tr("Query Layer"),
+                [QgsProcessing.TypeVectorPolygon, QgsProcessing.TypeVectorPoint]
             )
         )
 
         self.addParameter(
             QgsProcessingParameterEnum(
                 self.PARAMETER,
-                self.tr("Parameter"),
+                self.tr("Aggregation Type"),
                 options=self.parameters,
                 defaultValue=0,
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterString(
+                self.PERIOD, "Period (ISO 8601)", defaultValue="/P1M"
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterBoolean(
+                self.group_by_boundary,
+                self.tr("Group by Boundary"),
+                defaultValue=False,
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterDateTime(
+                self.date_start, "Start Date", defaultValue=start_date_string
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterDateTime(
+                self.date_end, "End Date", defaultValue=end_date_string
             )
         )
 
@@ -152,99 +229,10 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
         )
 
         self.addParameter(
-            QgsProcessingParameterBoolean(
-                self.check_activate_temporal,
-                self.tr("Qgis temporal feature"),
-                defaultValue=True,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterBoolean(
-                self.check_show_metadata,
-                self.tr("Show metadata"),
-                defaultValue=False,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterNumber(
-                self.timeout_input,
-                "Timeout",
-                type=QgsProcessingParameterNumber.Integer,
-                defaultValue=60,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterBoolean(
-                self.check_clip_geometry,
-                self.tr("Clip geometry"),
-                defaultValue=True,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterBoolean(
-                self.property_groups_check_tags,
-                self.tr("Tags"),
-                defaultValue=True,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterBoolean(
-                self.property_groups_check_metadata,
-                self.tr("Metadata"),
-                defaultValue=False,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterEnum(
-                self.data_aggregation_format,
-                self.tr("Format"),
-                options=self.formats,
-                defaultValue=0,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterDateTime(
-                self.date_start, "Start Date", defaultValue="2007-10-08"
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterDateTime(
-                self.date_end, "End Date", defaultValue="2023-07-28"
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterNumber(
-                self.YEARS,
-                "Years",
-                type=QgsProcessingParameterNumber.Integer,
-                defaultValue=0,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterNumber(
-                self.MONTHS,
-                "Months",
-                type=QgsProcessingParameterNumber.Integer,
-                defaultValue=0,
-            )
-        )
-
-        self.addParameter(
-            QgsProcessingParameterNumber(
-                self.DAYS,
-                "Days",
-                type=QgsProcessingParameterNumber.Integer,
-                defaultValue=1,
+            QgsProcessingParameterString(
+                self.FILTER_2,
+                self.tr("Filter 2"),
+                defaultValue="natural=tree",
             )
         )
 
@@ -253,7 +241,59 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
                 self.RADIUS,
                 "Radius [m]",
                 type=QgsProcessingParameterNumber.Integer,
-                defaultValue=100,
+                defaultValue=1000,
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterEnum(
+                self.GROUPBY,
+                self.tr("Group By"),
+                options=self.group_by,
+                defaultValue=0,
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterNumber(
+                self.timeout_input,
+                "Timeout",
+                type=QgsProcessingParameterNumber.Integer,
+                defaultValue=0,
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterEnum(
+                self.data_aggregation_format,
+                self.tr("Output Format"),
+                options=self.formats,
+                defaultValue=0,
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterString(
+                self.group_by_values_line_edit,
+                self.tr("Group by Values"),
+                optional=True,
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterString(
+                self.group_by_key_line_edit,
+                self.tr("Group by Key"),
+                optional=True,
+            )
+        )
+
+        # Qgis internal
+        self.addParameter(
+            QgsProcessingParameterBoolean(
+                self.check_show_metadata,
+                self.tr("Show metadata"),
+                defaultValue=False,
             )
         )
 
@@ -274,20 +314,20 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
         )
 
         self.addParameter(
-            QgsProcessingParameterString(
-                self.group_by_values_line_edit,
-                self.tr("Group by Values"),
-                optional=True,
+            QgsProcessingParameterBoolean(
+                self.check_activate_temporal,
+                self.tr("Qgis temporal feature"),
+                defaultValue=True,
             )
         )
 
         self.addParameter(
-            QgsProcessingParameterString(
-                self.group_by_key_line_edit,
-                self.tr("Group by Key"),
-                optional=True,
+            QgsProcessingParameterBoolean(
+                self.check_clip_geometry,
+                self.tr("Clip geometry"),
+                defaultValue=True,
             )
-        )
+        ),
 
     def processAlgorithm(self, parameters, context, feedback):
         """
@@ -299,57 +339,52 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
             geom = 1
         elif layer.geometryType() == QgsWkbTypes.PolygonGeometry:
             geom = 2
+
+        if self.parameterAsBool(parameters, self.group_by_boundary, context):
+            group_by_boundary_var = "/groupBy/boundary"
         else:
-            # implement user information
-            pass
+            group_by_boundary_var = ""
+
+        preference = self.parameters[
+            self.parameterAsInt(parameters, self.PARAMETER, context)
+        ]
+        groupBy = self.group_by[
+            self.parameterAsInt(parameters, self.GROUPBY, context)
+        ]
 
         processingParams = {
             "geom": geom,
             "selection": "data-Aggregation",
-            "preference": "contributions/count",
+            "preference": f"elements/{preference}/ratio{group_by_boundary_var}",
+            "preference_specification": groupBy,
             "filter": self.parameterAsString(parameters, self.FILTER, context),
-            "preference_specification": self.parameters[
-                self.parameterAsInt(parameters, self.PARAMETER, context)
-            ],
             "LAYER": self.parameterAsLayer(parameters, self.LAYER, context),
             "RADIUS": self.parameterAsInt(parameters, self.RADIUS, context),
-            "check_activate_temporal": self.parameterAsBool(
-                parameters, self.check_activate_temporal, context
-            ),
-            "check_show_metadata": self.parameterAsBool(
-                parameters, self.check_show_metadata, context
-            ),
-            "timeout_input": self.parameterAsInt(
-                parameters, self.timeout_input, context
-            ),
-            "check_clip_geometry": self.parameterAsBool(
-                parameters, self.check_clip_geometry, context
-            ),
-            "property_groups_check_tags": self.parameterAsBool(
-                parameters, self.property_groups_check_tags, context
-            ),
-            "property_groups_check_metadata": self.parameterAsBool(
-                parameters, self.property_groups_check_metadata, context
-            ),
-            "data_aggregation_format": self.formats[
-                self.parameterAsInt(
-                    parameters, self.data_aggregation_format, context
-                )
-            ],
             "date_start": self.parameterAsDateTime(
                 parameters, self.date_start, context
             ),
             "date_end": self.parameterAsDateTime(
                 parameters, self.date_end, context
             ),
-            "YEARS": self.parameterAsInt(parameters, self.YEARS, context),
-            "MONTHS": self.parameterAsInt(parameters, self.MONTHS, context),
-            "DAYS": self.parameterAsInt(parameters, self.DAYS, context),
-            "check_keep_geometryless": self.parameterAsBool(
-                parameters, self.check_keep_geometryless, context
+            "timeout_input": self.parameterAsInt(
+                parameters, self.timeout_input, context
             ),
-            "check_merge_geometries": self.parameterAsBool(
-                parameters, self.check_merge_geometries, context
+            "preference_specification": "",
+            "data_aggregation_format": "json",
+            "check_show_metadata": self.parameterAsBool(
+                parameters, self.check_show_metadata, context
+            ),
+            "timeout_input": self.parameterAsInt(
+                parameters, self.timeout_input, context
+            ),
+
+            "data_aggregation_format": self.formats[
+                self.parameterAsInt(
+                    parameters, self.data_aggregation_format, context
+                )
+            ],
+            "check_clip_geometry": self.parameterAsBool(
+                parameters, self.check_clip_geometry, context
             ),
             "group_by_values_line_edit": self.parameterAsString(
                 parameters, self.group_by_values_line_edit, context
@@ -357,6 +392,8 @@ class ElementsPerimeter(QgsProcessingAlgorithm):
             "group_by_key_line_edit": self.parameterAsString(
                 parameters, self.group_by_key_line_edit, context
             ),
+            'filter_2': self.parameterAsString(parameters, self.FILTER_2, context),
+            "filter": self.parameterAsString(parameters, self.FILTER, context),
         }
 
         run_processing_alg(processingParams, feedback)
